refactor(web): extract useTextEditor hook and shared utilities
Extracted text editor logic into separate files to improve maintainability and testability:

- Created `use-text-editor.ts` with `useTextEditor()` hook containing all TipTap editor setup logic
- Created `text-editor-shared.tsx` with shared types, utilities, and UI components to avoid circular dependencies
- Simplified `text-editor.tsx` from ~600 lines to ~50 lines by delegating to the hook
- Updated imports in `editable-hint.tsx` to use shared utilities
- Updated tests to import from shared file

This refactoring reduces complexity while maintaining all existing functionality.
All 12 smoke tests passing.
